Bringing examples in line with Dirt default samples.
Dirt includes a directory of snare drums called `sn`. Whilst this
is occasionally used in examples, there are still references to a
snare directory called `sd` which will just lead to errors -
potentially confusing for new errors.

This updates the _Creating Patterns_ pages to only refer to snare
drum directories included as standard in Dirt.